Sockets refactor - Removed unused code - Renamed files - Split gameplay socket from chat socket, this avoids a circular dependency with game.js and messages.js - Adjusted code to conform to standards
