--- src/socket-message-types.json
@@ -1,12 +1,0 @@
-{
-	"routes": {
-		"/gameplay": {
-			"initial_data": 0,
-			"request_physics": 1,
-			"new_marble": 2,
-			"state": 3,
-			"notification": 4,
-			"finished_marble": 5
-		}
-	}
-}
